ocr/lib-crawl-python/requirements.txt
adium-theme-ubuntu==0.3.4
attrs==15.2.0
beautifulsoup4==4.4.1
blinker==1.3
bzr==2.7.0
bzr-builddeb==2.8.6
chardet==2.3.0
configobj==5.0.6
cryptography==1.2.3
cssselect==1.0.0
debtags==2.0
defer==1.0.6
defusedxml==0.4.1
dirspec==13.10
distro-info===0.14build1
dnspython==1.12.0
duplicity==0.7.6
ecdsa==0.13
enum34==1.1.2
eyeD3==0.6.18
gyp==0.1
html5lib==0.999
httplib2==0.9.1
idna==2.0
ipaddress==1.0.16
keyring==7.3
launchpadlib==1.10.3
lazr.restfulclient==0.13.4
lazr.uri==1.0.3
lockfile==0.12.2
lxml==3.5.0
lz4==0.7.0
meld==3.14.2
mercurial==3.7.3
mutagen===1.31.-1
mysql-connector-python==2.0.4
mysql-utilities==1.6.1
ndg-httpsclient==0.4.0
netifaces==0.10.4
oauth==1.0.1
oauthlib==1.0.3
oneconf==0.3.9
PAM==0.4.2
paramiko==1.16.0
parsel==1.0.3
pexpect==4.0.1
Pillow==3.1.2
piston-mini-client==0.7.5
playmusicdecrypter==1.0
ptyprocess==0.5
pyasn1==0.1.9
pyasn1-modules==0.0.7
PyBluez==0.18
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0
PyDispatcher==2.0.5
pydns==2.3.6
pygobject==3.20.0
pygpgme==0.3
PyJWT==1.3.0
pykerberos==1.1.5
pyliblzma==0.5.3
pyodbc==3.0.7
PyOpenGL==3.0.2
pyOpenSSL==0.15.1
pyserial==3.0.1
pysqlite==2.7.0
python-apt==1.1.0b1
python-debian==0.1.27
python-debianbts==2.6.0
python-lzo==1.8
pyxdg==0.25
queuelib==1.4.2
rencode==1.0.3
reportbug===6.6.6ubuntu1
requests==2.9.1
scour==0.32
Scrapy==1.2.1
SecretStorage==2.1.3
service-identity==16.0.0
simplejson==3.8.1
six==1.10.0
SOAPpy==0.12.22
software-center-aptd-plugins==0.0.0
tails-installer==3.11.8
touchpad-indicator==1.1.0
Twisted==16.0.0
ubuntu-dev-tools===0.155ubuntu2
ubuntuone-client-data==14.4
unity-lens-photos==1.0
urlgrabber==3.9.1
urllib3==1.13.1
vboxapi==1.0
virtualenv==15.0.3
w3lib==1.16.0
wadllib==1.3.2
wstools==0.4.3
xpra==0.15.8
zope.interface==4.1.3
